+++ android/src/main/java/com/backendless/backendless_sdk/call_handlers/MessagingCallHandler.java
@@ -160,6 +160,8 @@ public class MessagingCallHandler implements MethodChannel.MethodCallHandler {
                 } else {
                     Backendless.Messaging.publish(channelName, message, publishOptions, callback);
                 }
+            } else if (deliveryOptions != null) {
+                Backendless.Messaging.publish(channelName, message, null, deliveryOptions, callback);
             } else {
                 Backendless.Messaging.publish(channelName, message, callback);
             }
@@ -170,6 +172,8 @@ public class MessagingCallHandler implements MethodChannel.MethodCallHandler {
                 } else {
                     Backendless.Messaging.publish(message, publishOptions, callback);
                 }
+            } else if (deliveryOptions != null) {
+                Backendless.Messaging.publish(message, null, deliveryOptions, callback);
             } else {
                 Backendless.Messaging.publish(message, callback);
             }
@@ -326,6 +330,15 @@ public class MessagingCallHandler implements MethodChannel.MethodCallHandler {
                     channel.addMessageListener(messageInfoCallback);
                 }
                 messageCallbacks.put(messageHandle, messageInfoCallback);
+                break;
+            case "Map":
+                EventAsyncCallback<HashMap> mapCallback = new EventAsyncCallback<>("Message", messageHandle);
+                if (selector != null) {
+                    channel.addMessageListener(selector, mapCallback, HashMap.class);
+                } else {
+                    channel.addMessageListener(mapCallback, HashMap.class);
+                }
+                messageCallbacks.put(messageHandle, mapCallback);
                 break;
             default:
                 throw new IllegalArgumentException("Custom type messages are unsupported for now");