+++ android/src/main/java/com/backendless/backendless_sdk/call_handlers/UserServiceCallHandler.java
@@ -7,6 +7,7 @@ import com.backendless.backendless_sdk.utils.FlutterCallback;
 import com.backendless.property.UserProperty;
 
 import java.util.List;
+import java.util.Map;
 
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
@@ -61,6 +62,15 @@ public class UserServiceCallHandler implements MethodChannel.MethodCallHandler {
                 break;
             case "Backendless.UserService.setUserToken":
                 setUserToken(call, result);
+                break;
+            case "Backendless.UserService.loginWithFacebook":
+                loginWithFacebook(call, result);
+                break;
+            case "Backendless.UserService.loginWithTwitter":
+                loginWithTwitter(call, result);
+                break;
+            case "Backendless.UserService.loginWithGoogle":
+                loginWithGoogle(call, result);
                 break;
             default:
                 result.notImplemented();
@@ -138,5 +148,66 @@ public class UserServiceCallHandler implements MethodChannel.MethodCallHandler {
         String userToken = call.argument("userToken");
         HeadersManager.getInstance().addHeader(HeadersManager.HeadersEnum.USER_TOKEN_KEY, userToken);
         result.success(null);
+    }
+
+    private void loginWithFacebook(MethodCall call, MethodChannel.Result result) {
+        String accessToken = call.argument("accessToken");
+        Map<String, String> fieldsMapping = call.argument("fieldsMapping");
+        Boolean stayLoggedIn = call.argument("stayLoggedIn");
+
+        FlutterCallback<BackendlessUser> callback = new FlutterCallback<>(result);
+
+        if (fieldsMapping != null) {
+            if (stayLoggedIn != null) {
+                Backendless.UserService.loginWithFacebookSdk(accessToken, fieldsMapping, callback, stayLoggedIn);
+            } else {
+                Backendless.UserService.loginWithFacebookSdk(accessToken, fieldsMapping, callback);
+            }
+        } else if (stayLoggedIn != null) {
+            Backendless.UserService.loginWithFacebookSdk(accessToken, callback, stayLoggedIn);
+        } else {
+            Backendless.UserService.loginWithFacebookSdk(accessToken, callback);
+        }
+    }
+
+    private void loginWithTwitter(MethodCall call, MethodChannel.Result result) {
+        String authToken = call.argument("authToken");
+        String authTokenSecret = call.argument("authTokenSecret");
+        Map<String, String> fieldsMapping = call.argument("fieldsMapping");
+        Boolean stayLoggedIn = call.argument("stayLoggedIn");
+
+        FlutterCallback<BackendlessUser> callback = new FlutterCallback<>(result);
+
+        if (fieldsMapping != null) {
+            if (stayLoggedIn != null) {
+                Backendless.UserService.loginWithTwitterSdk(authToken, authTokenSecret, fieldsMapping, callback, stayLoggedIn);
+            } else {
+                Backendless.UserService.loginWithTwitterSdk(authToken, authTokenSecret, fieldsMapping, callback);
+            }
+        } else if (stayLoggedIn != null) {
+            Backendless.UserService.loginWithTwitterSdk(authToken, authTokenSecret, callback, stayLoggedIn);
+        } else {
+            Backendless.UserService.loginWithTwitterSdk(authToken, authTokenSecret, callback);
+        }
+    }
+
+    private void loginWithGoogle(MethodCall call, MethodChannel.Result result) {
+        String accessToken = call.argument("accessToken");
+        Map<String, String> fieldsMapping = call.argument("fieldsMapping");
+        Boolean stayLoggedIn = call.argument("stayLoggedIn");
+
+        FlutterCallback<BackendlessUser> callback = new FlutterCallback<>(result);
+
+        if (fieldsMapping != null) {
+            if (stayLoggedIn != null) {
+                Backendless.UserService.loginWithGooglePlusSdk(accessToken, fieldsMapping, callback, stayLoggedIn);
+            } else {
+                Backendless.UserService.loginWithGooglePlusSdk(accessToken, fieldsMapping, callback);
+            }
+        } else if (stayLoggedIn != null) {
+            Backendless.UserService.loginWithGooglePlusSdk(accessToken, callback, stayLoggedIn);
+        } else {
+            Backendless.UserService.loginWithGooglePlusSdk(accessToken, callback);
+        }
     }
 }