+++ example/lib/test_table.dart
@@ -1,9 +1,0 @@
-import 'package:backendless_sdk/backendless_sdk.dart';
-
-@reflector
-class TestTable {
-  String foo;
-
-  @override
-  toString() => "TestTable: {foo: $foo}";
-}