+++ lib/backendless_sdk.dart
@@ -1,5 +1,6 @@
 library backendless_sdk;
 
+import 'dart:collection';
 import 'dart:io';
 import 'dart:typed_data';
 import 'dart:ui' show hashValues;
@@ -24,6 +25,8 @@ part 'src/modules/data/data_store.dart';
 part 'src/modules/data/property.dart';
 part 'src/modules/data/rt.dart';
 part 'src/modules/data/reflector.dart';
+part 'src/modules/data/map_to_property.dart';
+part 'src/modules/data/types.dart';
 part 'src/modules/events.dart';
 part 'src/modules/files/files.dart';
 part 'src/modules/files/file_info.dart';