+++ lib/src/modules/data/property.dart
@@ -3,7 +3,7 @@ part of backendless_sdk;
 abstract class AbstractProperty {
   String name;
   bool required;
-  DateTypeEnum type;
+  DataTypeEnum type;
 
   AbstractProperty({this.name, this.required = false, this.type});
 }
@@ -19,7 +19,7 @@ class ObjectProperty extends AbstractProperty {
     defaultValue = json['defaultValue'];
     name = json['name'];
     this.required = json['required'];
-    type = DateTypeEnum.values[json['type']];
+    type = DataTypeEnum.values[json['type']];
   }
 
   Map toJson() => {
@@ -29,9 +29,13 @@ class ObjectProperty extends AbstractProperty {
         'required': this.required,
         'type': type?.index,
       };
+
+  @override
+  String toString() =>
+      "ObjectProperty{name='$name', required=${this.required}, type=$type, relatedTable='$relatedTable', defaultValue=$defaultValue}";
 }
 
-enum DateTypeEnum {
+enum DataTypeEnum {
   UNKNOWN,
   INT,
   STRING,