+++ lib/src/modules/messaging/messaging.dart
@@ -1,9 +1,13 @@
 part of backendless_sdk;
 
+typedef void MessageHandler(Map<String, dynamic> message);
+
 class BackendlessMessaging {
   static const String DEFAULT_CHANNEL_NAME = "default";
-  static const MethodChannel _channel = const MethodChannel(
+  static const MethodChannel _messagingChannel = const MethodChannel(
       'backendless/messaging', StandardMethodCodec(BackendlessMessageCodec()));
+  static const MethodChannel _fcmServiceChannel =
+      const MethodChannel('backendless/messaging/push');
   static final Map<int, EventCallback> _joinCallbacks = <int, EventCallback>{};
   static final Map<int, EventCallback> _messageCallbacks =
       <int, EventCallback>{};
@@ -12,94 +16,58 @@ class BackendlessMessaging {
   static final Map<int, EventCallback> _userStatusCallbacks =
       <int, EventCallback>{};
   int _channelHandle = 0;
+  MessageHandler _onMessage;
 
   factory BackendlessMessaging() => _instance;
   static final BackendlessMessaging _instance =
       new BackendlessMessaging._internal();
 
   BackendlessMessaging._internal() {
-    _channel.setMethodCallHandler((MethodCall call) async {
-      if (call.method.contains("EventResponse")) {
-        Map<dynamic, dynamic> arguments = call.arguments;
-        int handle = arguments["handle"];
-        var response = arguments["response"];
-
-        switch (call.method) {
-          case ("Backendless.Messaging.Channel.Join.EventResponse"):
-            _joinCallbacks[handle].handleResponse();
-            break;
-          case ("Backendless.Messaging.Channel.Message.EventResponse"):
-            _messageCallbacks[handle].handleResponse(response);
-            break;
-          case ("Backendless.Messaging.Channel.Command.EventResponse"):
-            _commandCallbacks[handle].handleResponse(response);
-            break;
-          case ("Backendless.Messaging.Channel.UserStatus.EventResponse"):
-            _userStatusCallbacks[handle].handleResponse(response);
-            break;
-        }
-      } else if (call.method.contains("EventFault")) {
-        int handle = call.arguments["handle"];
-        String fault = call.arguments["fault"];
-
-        switch (call.method) {
-          case ("Backendless.Messaging.Channel.Join.EventFault"):
-            _joinCallbacks[handle].handleFault(fault);
-            break;
-          case ("Backendless.Messaging.Channel.Message.EventFault"):
-            _messageCallbacks[handle].handleFault(fault);
-            break;
-          case ("Backendless.Messaging.Channel.Command.EventFault"):
-            _commandCallbacks[handle].handleFault(fault);
-            break;
-          case ("Backendless.Messaging.Channel.UserStatus.EventFault"):
-            _userStatusCallbacks[handle].handleFault(fault);
-            break;
-        }
-      }
-    });
+    _messagingChannel.setMethodCallHandler(_handleMethod);
+    _fcmServiceChannel.setMethodCallHandler(_handleMethod);
   }
 
-  Future<MessageStatus> cancel(String messageId) => _channel.invokeMethod(
-      "Backendless.Messaging.cancel",
-      <String, dynamic>{"messageId": messageId});
-
-  Future<DeviceRegistration> getDeviceRegistration() =>
-      _channel.invokeMethod("Backendless.Messaging.getDeviceRegistration");
-
-  Future<MessageStatus> getMessageStatus(String messageId) =>
-      _channel.invokeMethod("Backendless.Messaging.getMessageStatus",
+  Future<MessageStatus> cancel(String messageId) =>
+      _messagingChannel.invokeMethod("Backendless.Messaging.cancel",
           <String, dynamic>{"messageId": messageId});
 
-  Future<MessageStatus> publish(Object message,
-      {String channelName,
-      PublishOptions publishOptions,
-      DeliveryOptions deliveryOptions}) {
-    if (deliveryOptions != null && publishOptions == null)
-      throw new ArgumentError(
-          "Argument 'deliveryOptions' should be defined with argument 'publishOptions'");
-    return _channel
-        .invokeMethod("Backendless.Messaging.publish", <String, dynamic>{
-      "message": message,
-      "channelName": channelName,
-      "publishOptions": publishOptions,
-      "deliveryOptions": deliveryOptions
-    });
-  }
+  Future<DeviceRegistration> getDeviceRegistration() => _messagingChannel
+      .invokeMethod("Backendless.Messaging.getDeviceRegistration");
+
+  Future<MessageStatus> getMessageStatus(String messageId) =>
+      _messagingChannel.invokeMethod("Backendless.Messaging.getMessageStatus",
+          <String, dynamic>{"messageId": messageId});
+
+  Future<MessageStatus> publish<T>(T message,
+          {String channelName,
+          PublishOptions publishOptions,
+          DeliveryOptions deliveryOptions}) =>
+      _messagingChannel
+          .invokeMethod("Backendless.Messaging.publish", <String, dynamic>{
+        "message": reflector.isCustomClass(message)
+            ? reflector.serialize(message)
+            : message,
+        "channelName": channelName,
+        "publishOptions": publishOptions,
+        "deliveryOptions": deliveryOptions
+      });
 
   Future<MessageStatus> pushWithTemplate(String templateName) =>
-      _channel.invokeMethod("Backendless.Messaging.pushWithTemplate",
+      _messagingChannel.invokeMethod("Backendless.Messaging.pushWithTemplate",
           <String, dynamic>{"templateName": templateName});
 
   Future<DeviceRegistrationResult> registerDevice(
-          [List<String> channels, DateTime expiration]) =>
-      _channel.invokeMethod("Backendless.Messaging.registerDevice",
-          <String, dynamic>{"channels": channels, "expiration": expiration});
+      [List<String> channels, DateTime expiration, MessageHandler onMessage]) {
+    _onMessage = onMessage;
+    return _messagingChannel.invokeMethod(
+        "Backendless.Messaging.registerDevice",
+        <String, dynamic>{"channels": channels, "expiration": expiration});
+  }
 
   Future<MessageStatus> sendEmail(
           String subject, BodyParts bodyParts, List<String> recipients,
           [List<String> attachments]) =>
-      _channel
+      _messagingChannel
           .invokeMethod("Backendless.Messaging.sendEmail", <String, dynamic>{
         "textMessage": bodyParts.textMessage,
         "htmlMessage": bodyParts.htmlMessage,
@@ -110,7 +78,7 @@ class BackendlessMessaging {
 
   Future<MessageStatus> sendHTMLEmail(
           String subject, String messageBody, List<String> recipients) =>
-      _channel.invokeMethod(
+      _messagingChannel.invokeMethod(
           "Backendless.Messaging.sendHTMLEmail", <String, dynamic>{
         "subject": subject,
         "messageBody": messageBody,
@@ -119,7 +87,7 @@ class BackendlessMessaging {
 
   Future<MessageStatus> sendTextEmail(
           String subject, String messageBody, List<String> recipients) =>
-      _channel.invokeMethod(
+      _messagingChannel.invokeMethod(
           "Backendless.Messaging.sendTextEmail", <String, dynamic>{
         "subject": subject,
         "messageBody": messageBody,
@@ -127,13 +95,13 @@ class BackendlessMessaging {
       });
 
   Future<int> unregisterDevice([List<String> channels]) =>
-      _channel.invokeMethod("Backendless.Messaging.unregisterDevice",
+      _messagingChannel.invokeMethod("Backendless.Messaging.unregisterDevice",
           <String, dynamic>{"channels": channels});
 
   Future<MessageStatus> sendEmailFromTemplate(
           String templateName, EmailEnvelope envelope,
           [Map<String, String> templateValues]) =>
-      _channel.invokeMethod(
+      _messagingChannel.invokeMethod(
           "Backendless.Messaging.sendEmailFromTemplate", <String, dynamic>{
         "templateName": templateName,
         "envelope": envelope,
@@ -142,10 +110,55 @@ class BackendlessMessaging {
 
   Future<Channel> subscribe([String channelName = DEFAULT_CHANNEL_NAME]) async {
     int handle = _channelHandle++;
-    return _channel.invokeMethod(
+    return _messagingChannel.invokeMethod(
         "Backendless.Messaging.subscribe", <String, dynamic>{
       "channelName": channelName,
       "channelHandle": handle
-    }).then((value) => new Channel(_channel, channelName, handle));
+    }).then((value) => new Channel(_messagingChannel, channelName, handle));
+  }
+
+  Future<dynamic> _handleMethod(MethodCall call) async {
+    if (call.method.contains("EventResponse")) {
+      Map<dynamic, dynamic> arguments = call.arguments;
+      int handle = arguments["handle"];
+      var response = arguments["response"];
+
+      switch (call.method) {
+        case ("Backendless.Messaging.Channel.Join.EventResponse"):
+          _joinCallbacks[handle].handleResponse();
+          break;
+        case ("Backendless.Messaging.Channel.Message.EventResponse"):
+          _messageCallbacks[handle].handleResponse(response);
+          break;
+        case ("Backendless.Messaging.Channel.Command.EventResponse"):
+          _commandCallbacks[handle].handleResponse(response);
+          break;
+        case ("Backendless.Messaging.Channel.UserStatus.EventResponse"):
+          _userStatusCallbacks[handle].handleResponse(response);
+          break;
+      }
+    } else if (call.method.contains("EventFault")) {
+      int handle = call.arguments["handle"];
+      String fault = call.arguments["fault"];
+
+      switch (call.method) {
+        case ("Backendless.Messaging.Channel.Join.EventFault"):
+          _joinCallbacks[handle].handleFault(fault);
+          break;
+        case ("Backendless.Messaging.Channel.Message.EventFault"):
+          _messageCallbacks[handle].handleFault(fault);
+          break;
+        case ("Backendless.Messaging.Channel.Command.EventFault"):
+          _commandCallbacks[handle].handleFault(fault);
+          break;
+        case ("Backendless.Messaging.Channel.UserStatus.EventFault"):
+          _userStatusCallbacks[handle].handleFault(fault);
+          break;
+      }
+    } else if (call.method == "onMessage") {
+      if (_onMessage != null) {
+        _onMessage(call.arguments.cast<String, dynamic>());
+      }
+    }
   }
 }